+++ R/cld-emm.R
@@ -70,32 +70,60 @@
 #'   If \code{TRUE}, the order of use of the letters is reversed.
 #'   In addition, if both \code{sort} and \code{reversed} are TRUE, the sort
 #'   order of results is reversed.
+#' @param signif.sets Logical value. If \code{FALSE} (and \code{delta = 0}), a 
+#'   \sQuote{traditional}
+#'   compact-letter display is constructed with groupings representing sets of
+#'   estimates that are not statistically different. If \code{TRUE}, the criteria
+#'   are reversed so that two estimates sharing the same symbol test as significantly
+#'   different. See also \code{delta}.
+#' @param delta Numeric value passed to \code{\link{test.emmGrid}}. If this
+#'   is positive, it is used as an equivalence threshold in the TOST procedure for
+#'   two-sided equivalence testing. In the resulting compact letter display,
+#'   two estimates share the same grouping letter only if they are found to be
+#'   statistically equivalent -- that is, groupings reflect actual \emph{findings}
+#'   of equivalence rather than failure to find a significant difference.
+#'   When \code{delta} is nonzero, \code{signif.sets} is ignored.
 #' @param ... Arguments passed to \code{\link{contrast}} (for example,
 #'   an \code{adjust} method)
 #' @references Piepho, Hans-Peter (2004) An algorithm for a letter-based 
 #'   representation of all pairwise comparisons, 
 #'   Journal of Computational and Graphical Statistics, 
 #'   13(2), 456-466.
+#' 
+#' @return
+#' A \code{\link{summary_emm}} object showing the estimated marginal means
+#' plus an additional column labeled \code{.group} (when \code{signif.sets = FALSE}), 
+#' \code{.signif.set} (when \code{signif.sets = TRUE}), or \code{.equiv.set} 
+#' (when \code{delta > 0}).
+#' 
+#' 
 #'   
 #' @note
-#' We warn that such displays encourage a poor
-#' practice in interpreting significance tests. CLDs are misleading because they
+#' We warn that the default display encourages a poor
+#' practice in interpreting significance tests. Such CLDs are misleading because they
 #' visually group means with comparisons \emph{P} > \code{alpha} as though they 
 #' are equal, when in fact we have only failed to prove that they differ.
-#' As alternatives, consider \code{\link{pwpp}} (graphical display of \emph{P} 
+#' A better alternative if one wants to show groupings is to specify an equivalence
+#' threshold \code{delta}; then groupings will be based on actual findings of
+#' equivalence. Another way to display actual findings is to set
+#' \code{signif.sets = TRUE}, so that estimates in the same group are those 
+#' found to be statistically \emph{different}. Obviously, these different options
+#' require different interpretations of the results; the annotations and the label
+#' given the final column help guide how to assess the results.
+#' 
+#' As further alternatives, consider \code{\link{pwpp}} (graphical display of \emph{P} 
 #' values) or \code{\link{pwpm}} (matrix display).
 #'
 #' @method cld emmGrid
 #' @examples
-#' if(requireNamespace("multcomp")) {
-#'     pigs.lm <- lm(log(conc) ~ source + factor(percent), data = pigs)
-#'     pigs.emm <- emmeans(pigs.lm, "percent", type = "response")
-#'     multcomp::cld(pigs.emm, alpha = 0.10, Letters = LETTERS)
-#' }
-cld.emmGrid = function(object, details=FALSE, sort=TRUE, 
-                       by, alpha=.05, 
+#' if(requireNamespace("multcomp"))
+#'     emm_example("cld-multcomp")
+#'     # Use emm_example("cld-multcomp", list = TRUE) # to just list the code
+#'
+cld.emmGrid = function(object, details = FALSE, sort = TRUE, 
+                       by, alpha = .05, 
                        Letters = c("1234567890",LETTERS,letters), 
-                       reversed=FALSE, ...) {
+                       reversed=FALSE, signif.sets = FALSE, delta = 0, ...) {
     if (!is.na(object@post.beta)[1]) {
         message("NOTE: Summary and groupings are based on frequentist results")
         object@post.beta = matrix(NA)
@@ -121,10 +149,18 @@ cld.emmGrid = function(object, details=FALSE, sort=TRUE,
     attr(emmtbl, "by.vars") = by
     object@misc$by.vars = by
     
-    prwise = contrast(object, "revpairwise", by=by)    
-    pwtbl = test(prwise, ...)
+    prwise = contrast(object, "revpairwise", by = by)
+    testargs = list(object = prwise, delta = delta, ...)
+    testargs$side = NULL   # can't be doin' 1-sided tests!
+    pwtbl = do.call(test, testargs)
     
-    p.boo = (pwtbl$p.value < alpha)
+    if (delta > 0)
+        signif.sets = TRUE
+    
+    p.boo = if(signif.sets)
+        (pwtbl$p.value >= alpha)
+    else
+        (pwtbl$p.value < alpha)
     if(is.null(by)) {
         by.rows = list(seq_len(nrow(pwtbl)))
         by.out = list(seq_len(nrow(emmtbl)))
@@ -178,13 +214,26 @@ cld.emmGrid = function(object, details=FALSE, sort=TRUE,
         r = by.out[[i]]
         emmtbl[r, ] = emmtbl[rev(r), ]
     }
-    
-    dontusemsg = paste0("NOTE: If two or more means share the same grouping letter,\n",
-                        "      then we cannot show them to be different.\n",
-                        "      But we also did not show them to be the same.")
-    
-    attr(emmtbl, "mesg") = c(attr(emmtbl,"mesg"), attr(pwtbl, "mesg"), 
-                             paste("significance level used: alpha =", alpha), dontusemsg)
+    if (!signif.sets) {
+        dontusemsg = paste0("NOTE: If two or more means share the same grouping symbol,\n",
+                            "      then we cannot show them to be different.\n",
+                            "      But we also did not show them to be the same.")
+        
+        attr(emmtbl, "mesg") = c(attr(emmtbl,"mesg"), attr(pwtbl, "mesg"), 
+                                 paste("significance level used: alpha =", alpha), dontusemsg)
+    }
+    else if (delta == 0) {
+        names(emmtbl)[names(emmtbl) == ".group"] = ".signif.set"
+        attr(emmtbl, "mesg") = attr(emmtbl, "mesg") = c(attr(emmtbl,"mesg"), attr(pwtbl, "mesg"), 
+            paste("significance level used: alpha =", alpha), 
+            "Estimates sharing the same symbol are significantly different")
+    }
+    else {
+        names(emmtbl)[names(emmtbl) == ".group"] = ".equiv.set"
+        attr(emmtbl, "mesg") = attr(emmtbl, "mesg") = c(attr(emmtbl,"mesg"), attr(pwtbl, "mesg"), 
+            paste("significance level used: alpha =", alpha), 
+            "Estimates sharing the same symbol test as equivalent")
+    }
     
     if (details)
         list(emmeans = emmtbl, comparisons = pwtbl)